requirements.txt
# Requisitos para Healthcare ETL + Telegram Bot
pandas>=2.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pyTelegramBotAPI>=4.0
nbconvert>=7.0
jupyter>=1.0